Move all database queries off the ui thread & add a ViewModel for MainActivity (#4786)
- Move all database queries off the ui thread - this is a massive
performance improvement
- ViewModel for MainActivity - this makes MainActivity smaller and
network requests won't be retried when rotating the screen
- removes the Push Notification Migration feature. We had it long
enough, all users who want push notifications should be migrated by now
- AccountEntity is now immutable
- converted BaseActivity to Kotlin
- The header image of Accounts is now cached as well

+++ app/src/main/java/com/keylesspalace/tusky/components/timeline/viewmodel/CachedTimelineViewModel.kt
@@ -86,9 +86,9 @@ class CachedTimelineViewModel @Inject constructor(
         config = PagingConfig(
             pageSize = LOAD_AT_ONCE
         ),
-        remoteMediator = CachedTimelineRemoteMediator(accountManager, api, db),
+        remoteMediator = CachedTimelineRemoteMediator(this, api, db),
         pagingSourceFactory = {
-            db.timelineDao().getHomeTimeline(account.id).also { newPagingSource ->
+            db.timelineDao().getHomeTimeline(accountId).also { newPagingSource ->
                 this.currentPagingSource = newPagingSource
             }
         }
@@ -118,27 +118,27 @@ class CachedTimelineViewModel @Inject constructor(
     override fun changeExpanded(expanded: Boolean, status: StatusViewData.Concrete) {
         viewModelScope.launch {
             db.timelineStatusDao()
-                .setExpanded(account.id, status.actionableId, expanded)
+                .setExpanded(accountId, status.actionableId, expanded)
         }
     }
 
     override fun changeContentShowing(isShowing: Boolean, status: StatusViewData.Concrete) {
         viewModelScope.launch {
             db.timelineStatusDao()
-                .setContentShowing(account.id, status.actionableId, isShowing)
+                .setContentShowing(accountId, status.actionableId, isShowing)
         }
     }
 
     override fun changeContentCollapsed(isCollapsed: Boolean, status: StatusViewData.Concrete) {
         viewModelScope.launch {
             db.timelineStatusDao()
-                .setContentCollapsed(account.id, status.actionableId, isCollapsed)
+                .setContentCollapsed(accountId, status.actionableId, isCollapsed)
         }
     }
 
     override fun clearWarning(status: StatusViewData.Concrete) {
         viewModelScope.launch {
-            db.timelineStatusDao().clearWarning(account.id, status.actionableId)
+            db.timelineStatusDao().clearWarning(accountId, status.actionableId)
         }
     }
 
@@ -154,12 +154,12 @@ class CachedTimelineViewModel @Inject constructor(
                 val accountDao = db.timelineAccountDao()
 
                 timelineDao.insertHomeTimelineItem(
-                    Placeholder(placeholderId, loading = true).toEntity(tuskyAccountId = account.id)
+                    Placeholder(placeholderId, loading = true).toEntity(tuskyAccountId = accountId)
                 )
 
                 val (idAbovePlaceholder, idBelowPlaceholder) = db.withTransaction {
-                    timelineDao.getIdAbove(account.id, placeholderId) to
-                        timelineDao.getIdBelow(account.id, placeholderId)
+                    timelineDao.getIdAbove(accountId, placeholderId) to
+                        timelineDao.getIdBelow(accountId, placeholderId)
                 }
                 val response = when (readingOrder) {
                     // Using minId, loads up to LOAD_AT_ONCE statuses with IDs immediately
@@ -184,12 +184,17 @@ class CachedTimelineViewModel @Inject constructor(
                     return@launch
                 }
 
+                val account = activeAccountFlow.value
+                if (account == null) {
+                    return@launch
+                }
+
                 db.withTransaction {
-                    timelineDao.deleteHomeTimelineItem(account.id, placeholderId)
+                    timelineDao.deleteHomeTimelineItem(accountId, placeholderId)
 
                     val overlappedStatuses = if (statuses.isNotEmpty()) {
                         timelineDao.deleteRange(
-                            account.id,
+                            accountId,
                             statuses.last().id,
                             statuses.first().id
                         )
@@ -198,14 +203,14 @@ class CachedTimelineViewModel @Inject constructor(
                     }
 
                     for (status in statuses) {
-                        accountDao.insert(status.account.toEntity(account.id))
-                        status.reblog?.account?.toEntity(account.id)
+                        accountDao.insert(status.account.toEntity(accountId))
+                        status.reblog?.account?.toEntity(accountId)
                             ?.let { rebloggedAccount ->
                                 accountDao.insert(rebloggedAccount)
                             }
                         statusDao.insert(
                             status.actionableStatus.toEntity(
-                                tuskyAccountId = account.id,
+                                tuskyAccountId = accountId,
                                 expanded = account.alwaysOpenSpoiler,
                                 contentShowing = account.alwaysShowSensitiveMedia || !status.actionableStatus.sensitive,
                                 contentCollapsed = true
@@ -213,7 +218,7 @@ class CachedTimelineViewModel @Inject constructor(
                         )
                         timelineDao.insertHomeTimelineItem(
                             HomeTimelineEntity(
-                                tuskyAccountId = account.id,
+                                tuskyAccountId = accountId,
                                 id = status.id,
                                 statusId = status.actionableId,
                                 reblogAccountId = if (status.reblog != null) {
@@ -239,7 +244,7 @@ class CachedTimelineViewModel @Inject constructor(
                             Placeholder(
                                 idToConvert,
                                 loading = false
-                            ).toEntity(account.id)
+                            ).toEntity(accountId)
                         )
                     }
                 }
@@ -252,7 +257,7 @@ class CachedTimelineViewModel @Inject constructor(
     }
 
     private suspend fun loadMoreFailed(placeholderId: String, e: Exception) {
-        Log.w("CachedTimelineVM", "failed loading statuses", e)
+        Log.w(TAG, "failed loading statuses", e)
         val activeAccount = accountManager.activeAccount!!
         db.timelineDao()
             .insertHomeTimelineItem(Placeholder(placeholderId, loading = false).toEntity(activeAccount.id))
@@ -266,16 +271,19 @@ class CachedTimelineViewModel @Inject constructor(
     }
 
     override fun saveReadingPosition(statusId: String) {
-        accountManager.activeAccount?.let { account ->
-            Log.d(TAG, "Saving position at: $statusId")
-            account.lastVisibleHomeTimelineStatusId = statusId
-            accountManager.saveAccount(account)
+        viewModelScope.launch {
+            accountManager.activeAccount?.let { account ->
+                Log.d(TAG, "Saving position at: $statusId")
+                accountManager.updateAccount(account) {
+                    copy(lastVisibleHomeTimelineStatusId = statusId)
+                }
+            }
         }
     }
 
     override suspend fun invalidate() {
         // invalidating when we don't have statuses yet can cause empty timelines because it cancels the network load
-        if (db.timelineDao().getHomeTimelineItemCount(account.id) > 0) {
+        if (db.timelineDao().getHomeTimelineItemCount(accountId) > 0) {
             currentPagingSource?.invalidate()
         }
     }
